Make the XML well-formed again.
I don't know why there were two electronconfig2="" attributes for the
different spectra, I suspect that is a bug, so I made the first such attribute
read electronconfig1="".

Also, some discoverer names had the appropriate entities added, and the
"degrees" symbol isn't a part of XML so I defined a degrees entity and
replaced all occurrences of the symbol with the entity.

svn path=/trunk/KDE/kdeedu/kalzium/; revision=435673