Ok, here we are. This is the new painting system for Kalzium.
Almost all the "pieces" of the periodic table are splitted in small classes
built in a sort of plugin form, with factories to easily get them.
The central class is the KalziumPainter, which encapsulate a QPainter denying
the direct access to it. It's the KalziumPainter that queries the various
"plugin" classes for information about the table and its painting.

All the system seems working quite fine, just sare left to implements some of
the old schemes and gradiets, and some small bugs that hopefully will be fixed
soon.

svn path=/trunk/KDE/kdeedu/kalzium/; revision=474144