Make the label dropdown work. Element names are taken from OpenBabel, no i18n possible here. Kalzium has i18n element names but linking to that code from libavogadro is impractical and not future-proof (think when we'll use libavogadro as an external dependency).
I had to add quick-and-dirty hooks to the libavogadro API so this is 
probably not for merging into avo trunk.

Also add a quick way to get the active rendering engine so that 
retrieving the radius of an atom can be done at a low cost. This also 
isn't mergeable upstream (already discussed).


svn path=/trunk/KDE/kdeedu/kalzium/; revision=666905