Update min and max value at the same time in order to be more friendly with typing in the min/max box
REVIEW: 114006

This patch makes the behavior while typing in the min/max boxes more friendly.

Currently if one types in the maximum value box, the minimum value will stuck at a small value due to the small value in the maximum value box before one finish typing. (I don't really call this a bug, but I think it is not friendly.)

This patch tries to update both the maximum and minimum values whenever one of them changes (in the right order to always respect the one that changes) so that the actual range displayed is always the closed possible to the input values. It also covers the cases when min/max values change from illegal values to legal values (max > min) in some other ways.