Correct formatting of year of discovery
The year of discovery was formatted as an amount,
instead of a number, so instead of being displayed
as, e.g., ‘1961’, it was being displayed as
‘1,961’, ‘1.961’ or ‘1 961’, depending on the locale.

The old KUIT tag ‘<numid>’ was supposed to fix the
formatting, but is no longer supported. The recommended
fix is to use QString::number(), as documented in:
https://api.kde.org/frameworks/ki18n/html/prg_guide.html#subs_notes

The ‘<numid>’ tags have not be removed from the string,
as that would break string freeze in the stable branch.
I will remove them in a separate commit in ‘master’.