Removed overlay symbol text in iconic view
Iconic view is a bit of a mess. One of the reasons is there is duplicate text of the symbol. One baked in the icon itself, and another overlay text which is drawn. 

The overlay text is completely useless in iconic view, in fact it clutters the icons and you can barely see anything. Additionally the proton number is also removed in this MR, as half of the proton numbers are not visible as they are white on top of white icons. In my opinion, iconic view is not used to get the quick view of the numbers and they can be omitted in this case without losing any important information. There are other schemes that are more friendly towards quick info requirements. Iconic view should be focus on the icons, and nothing else.

To prevent the drawing of this overlay text, I added a simple if statement in the _ElementItem:paint_ which checks if the current scheme is iconic view. An interesting thing is there is a second overlay being drawn in a completely different file "_kalziumdataobject_" while loading the icon set. This text is usually covered by the one from _ElementItem_, but becomes visible once that one is turned off. I do not know the purpose of this, but i commented it out instead of deleting just to be safe.