Forbid any gradient for iconic view
This is another MR to try to improve the iconic view, or in this case limit it.

Iconic view does not work with any gradient. It does not have any gradient applied, and even if it did, the icons would make it impossible to see the gradient. Therefore, the best option is to forbid the possibility of viewing the iconic view with gradient. Because now, its just confusing.

This implementation does a simple thing. Whenever the iconic view is selected, if the user selects any gradient, the scheme defaults to Monochrome instead of iconic. And if any gradient is selected, selecting the iconic view, defaults to "None" gradient.

This was done by doing a simple check in the _Kalzium::slotSwitchtoLookScheme_ and _Kalzium::slotSwitchtoLookGradient_.